requirements.txt
Django==1.11.5
django-filter==1.0.4
djangorestframework==3.6.4
mysqlclient==1.3.12
psycopg2==2.7.3.1
pytz==2017.2
